refactor:(#8)Marker 컴포넌트 분리
