+++ src/components/maps/JobMarker.jsx
@@ -1,0 +1,16 @@
+import { MapMarker } from 'react-kakao-maps-sdk';
+
+const JobMarker = ({ job, onClick }) => (
+  <MapMarker
+    key={job.id}
+    position={{ lat: job.lat, lng: job.lng }}
+    clickable={true}
+    onClick={() => onClick(job.id)}
+    image={{
+      src: '/public/images/PINGER_marker.png',
+      size: { width: 30, height: 30 },
+    }}
+  />
+);
+
+export default JobMarker;
